// src/generated/models/task-model.ts
export const TaskStatusKeyToLabel = {
  'StatusKey0': 'Pass',
  'StatusKey1': 'Fail',
  'StatusKey2': 'In Progress'
} as const;
export type TaskStatusKey = keyof typeof TaskStatusKeyToLabel;

export interface Task {
  /**
   * @displayName Task
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Task Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  taskname: string;
  /**
   * @displayName Adv Latency (s)
   */
  advlatencys?: number;
  /**
   * @displayName Dialogue Turns
   */
  dialogueturns?: number;
  /**
   * @displayName Status
   */
  statusKey?: TaskStatusKey;
}

export const _Task = 'Task' as const;